+++ packages/vit-s/src/VitS.js
@@ -23,7 +23,7 @@ import {
 } from './view-config-utils.js';
 
 function logConfig(config, name) {
-  console.groupCollapsed(`🚄 Vitessce (${META_VERSION.version}) ${name}`);
+  console.groupCollapsed(`🚄 VitS (${META_VERSION.version}) ${name}`);
   console.info(`data:,${JSON.stringify(config)}`);
   console.info(JSON.stringify(config, null, 2));
   console.groupEnd();
@@ -117,7 +117,7 @@ export function VitS(props) {
     const result = latestConfigSchema.safeParse(config);
     if (result.success) {
       const upgradedConfig = result.data;
-      logConfig(upgradedConfig, 'upgraded view config');
+      logConfig(upgradedConfig, 'parsed view config');
       // Perform second round of parsing against plugin-specific config schema.
       const pluginSpecificResult = pluginSpecificConfigSchema.safeParse(upgradedConfig);
       // Initialize the view config according to the initStrategy.